feat: converted data category to tsx
