--- src/prefabs/dataTableColumn.js
@@ -1,159 +1,0 @@
-(() => ({
-  name: 'Data Table Column',
-  icon: 'DataTableColumn',
-  category: 'DATA',
-  keywords: ['Data', 'table', 'column', 'datatablecolumn'],
-  beforeCreate: ({
-    components: { Content, Header, Field, Footer, PropertySelector },
-    prefab,
-    save,
-    close,
-  }) => {
-    const [property, setProperty] = React.useState('');
-
-    return (
-      <>
-        <Header
-          onClose={close}
-          title="New column"
-          subtitle="You can choose a property for your datatable here or choose to add it later by clicking 'add without configuration'"
-        />
-        <Content>
-          <Field label="Property">
-            <PropertySelector
-              onChange={(value) => {
-                setProperty(value);
-              }}
-              value={property}
-            />
-          </Field>
-        </Content>
-        <Footer
-          onSave={() => {
-            const newPrefab = { ...prefab };
-
-            newPrefab.structure[0].options[1].value = property;
-            save(newPrefab);
-          }}
-          onSkip={() => {
-            const newPrefab = { ...prefab };
-            save(newPrefab);
-          }}
-          onClose={close}
-        />
-      </>
-    );
-  },
-  structure: [
-    {
-      name: 'DataTableColumn',
-      options: [
-        {
-          value: true,
-          label: 'Initial visibility',
-          key: 'visible',
-          type: 'TOGGLE',
-          configuration: {
-            as: 'VISIBILITY',
-          },
-        },
-        {
-          value: '',
-          label: 'Property',
-          key: 'property',
-          type: 'PROPERTY',
-        },
-        {
-          type: 'TOGGLE',
-          label: 'Sortable',
-          key: 'sortable',
-          value: false,
-        },
-        {
-          type: 'VARIABLE',
-          label: 'Header text',
-          key: 'headerText',
-          value: [''],
-        },
-        {
-          value: 'Body1',
-          label: 'Header Type',
-          key: 'type',
-          type: 'FONT',
-        },
-        {
-          type: 'VARIABLE',
-          label: 'Content',
-          key: 'content',
-          value: [''],
-          configuration: {
-            as: 'MULTILINE',
-          },
-        },
-        {
-          value: 'Body1',
-          label: 'Body type',
-          key: 'bodyType',
-          type: 'FONT',
-        },
-        {
-          type: 'CUSTOM',
-          label: 'Column Alignment',
-          key: 'horizontalAlignment',
-          value: 'left',
-          configuration: {
-            as: 'BUTTONGROUP',
-            dataType: 'string',
-            allowedInput: [
-              { name: 'Left', value: 'left' },
-              { name: 'Center', value: 'center' },
-              { name: 'Right', value: 'right' },
-            ],
-          },
-        },
-        {
-          type: 'SIZE',
-          label: 'Width',
-          key: 'width',
-          value: '',
-          configuration: {
-            as: 'UNIT',
-          },
-        },
-        {
-          type: 'COLOR',
-          label: 'Background',
-          key: 'background',
-          value: 'Transparent',
-        },
-        {
-          type: 'COLOR',
-          label: 'Border color',
-          key: 'borderColor',
-          value: 'Light',
-        },
-        {
-          value: false,
-          label: 'Advanced settings',
-          key: 'advancedSettings',
-          type: 'TOGGLE',
-        },
-        {
-          type: 'VARIABLE',
-          label: 'Test attribute',
-          key: 'dataComponentAttribute',
-          value: ['DataTableColumn'],
-          configuration: {
-            condition: {
-              type: 'SHOW',
-              option: 'advancedSettings',
-              comparator: 'EQ',
-              value: true,
-            },
-          },
-        },
-      ],
-      descendants: [],
-    },
-  ],
-}))();
