+++ src/prefabs/detailView.tsx
@@ -12,10 +12,7 @@ import {
 } from '@betty-blocks/component-sdk';
 import { Box } from './structures/Box';
 import { options as boxOptions } from './structures/Box/options';
-import { Column } from './structures/Column';
-import { options as columnOptions } from './structures/Column/options';
 import { DataContainer } from './structures/DataContainer';
-import { DetailViewChild } from './structures/DetailViewChild';
 import { Row } from './structures/Row';
 import { options as rowOptions } from './structures/Row/options';
 
@@ -36,7 +33,15 @@ const beforeCreate = ({
   close,
 }: BeforeCreateArgs) => {
   const [modelId, setModelId] = React.useState('');
-  const [properties, setProperties] = React.useState([]);
+  const [properties, setProperties] = React.useState([
+    // {
+    //   defaultValue: '',
+    //   id: [''],
+    //   kind: '',
+    //   label: '',
+    //   type: '',
+    // },
+  ]);
   const [numberOfColumns, setNumberOfColumns] = React.useState('6');
   const sideBySide = false;
 
@@ -45,18 +50,18 @@ const beforeCreate = ({
   });
   const [errorMessage, setErrorMessage] = React.useState('');
 
-  // const enrichVarObj = (obj: any) => {
-  //   const returnObj = obj;
-  //   if (data && data.model) {
-  //     const property = data.model.properties.find(
-  //       (prop: any) => prop.id === returnObj.id[0],
-  //     );
-  //     if (property) {
-  //       returnObj.name = `{{ ${data.model.name}.${property.name} }}`;
-  //     }
-  //   }
-  //   return returnObj;
-  // };
+  const enrichVarObj = (obj: any) => {
+    const returnObj = obj;
+    if (data && data.model) {
+      const property = data.model.properties.find(
+        (prop: any) => prop.id === returnObj.id[0],
+      );
+      if (property) {
+        returnObj.name = `{{ ${data.model.name}.${property.name} }}`;
+      }
+    }
+    return returnObj;
+  };
 
   const getDescendantByRef = (refValue: string, structure: any) =>
     structure.reduce((acc: string, component: PrefabReference) => {
@@ -136,7 +141,6 @@ const beforeCreate = ({
           newPrefab.variables.push({
             kind: 'integer',
             name: variableName,
-            // pageId: pageUuid,
             ref: {
               id: '#idVariable',
             },
@@ -145,36 +149,1354 @@ const beforeCreate = ({
             (o) => o.key === 'model',
           ) as ValueDefault;
           modelOption.value = modelId;
-
-          // TODO: for each modelProperty --> Column + DetailViewChild with said property
-          if (!sideBySide) {
-            // const boxStructure = getDescendantByRef(
-            //   '#box',
-            //   structure.descendants,
-            // );
-            // const detailViewStructure = getDescendantByRef(
-            //   '#detailViewChild',
-            //   structure.descendants,
-            // );
-            // const insertedData = boxStructure.descendants.pop();
-            // insertedData.descendants[0].options[1].value = numberOfColumns;
-            // insertedData.descendants[0].options[2].value = numberOfColumns;
-            // debugger;
-            // insertedData.descendants[0].descendants = [];
-            // properties.forEach((property) => {
-            //   // descendants.push(Column({}, [DetailViewChild({})])), but that doesn't work yet
-            //   detailViewStructure.options[0].value =
-            //     property.kind === 'IMAGE' ? '' : enrichVarObj(property);
-            //   detailViewStructure.options[1].value =
-            //     property.kind === 'IMAGE' ? [`${property.label}`] : [];
-
-            //   insertedData.descendants[0].descendants.push(detailViewStructure);
-            // });
-            if (structure.descendants[0].type === 'COMPONENT') {
-              // structure.descendants[0].descendants.push(insertedData);
-              save(newPrefab);
+          if (
+            newPrefab.structure[0].type === 'COMPONENT' &&
+            newPrefab.structure[0].descendants[0].type === 'COMPONENT' &&
+            newPrefab.structure[0].descendants[0].descendants[0].type ===
+              'COMPONENT'
+          ) {
+            if (!sideBySide) {
+              properties.forEach((property) => {
+                if (
+                  newPrefab.structure[0].type === 'COMPONENT' &&
+                  newPrefab.structure[0].descendants[0].type === 'COMPONENT' &&
+                  newPrefab.structure[0].descendants[0].descendants[0].type ===
+                    'COMPONENT'
+                ) {
+                  newPrefab.structure[0].descendants[0].descendants[0].descendants.push(
+                    {
+                      name: 'Column',
+                      options: [
+                        {
+                          label: 'Toggle visibility',
+                          key: 'visible',
+                          value: true,
+                          type: 'TOGGLE',
+                          configuration: {
+                            as: 'VISIBILITY',
+                          },
+                        },
+                        {
+                          value: numberOfColumns,
+                          label: 'Column width',
+                          key: 'columnWidth',
+                          type: 'CUSTOM',
+                          configuration: {
+                            as: 'DROPDOWN',
+                            dataType: 'string',
+                            allowedInput: [
+                              { name: 'Fit content', value: 'fitContent' },
+                              { name: 'Flexible', value: 'flexible' },
+                              { name: 'Hidden', value: 'hidden' },
+                              { name: '1', value: '1' },
+                              { name: '2', value: '2' },
+                              { name: '3', value: '3' },
+                              { name: '4', value: '4' },
+                              { name: '5', value: '5' },
+                              { name: '6', value: '6' },
+                              { name: '7', value: '7' },
+                              { name: '8', value: '8' },
+                              { name: '9', value: '9' },
+                              { name: '10', value: '10' },
+                              { name: '11', value: '11' },
+                              { name: '12', value: '12' },
+                            ],
+                          },
+                        },
+                        {
+                          value: numberOfColumns,
+                          label: 'Column width (tablet landscape)',
+                          key: 'columnWidthTabletLandscape',
+                          type: 'CUSTOM',
+                          configuration: {
+                            as: 'DROPDOWN',
+                            dataType: 'string',
+                            allowedInput: [
+                              { name: 'Fit content', value: 'fitContent' },
+                              { name: 'Flexible', value: 'flexible' },
+                              { name: 'Hidden', value: 'hidden' },
+                              { name: '1', value: '1' },
+                              { name: '2', value: '2' },
+                              { name: '3', value: '3' },
+                              { name: '4', value: '4' },
+                              { name: '5', value: '5' },
+                              { name: '6', value: '6' },
+                              { name: '7', value: '7' },
+                              { name: '8', value: '8' },
+                              { name: '9', value: '9' },
+                              { name: '10', value: '10' },
+                              { name: '11', value: '11' },
+                              { name: '12', value: '12' },
+                            ],
+                          },
+                        },
+                        {
+                          value: 'flexible',
+                          label: 'Column width (tablet portrait)',
+                          key: 'columnWidthTabletPortrait',
+                          type: 'CUSTOM',
+                          configuration: {
+                            as: 'DROPDOWN',
+                            dataType: 'string',
+                            allowedInput: [
+                              { name: 'Fit content', value: 'fitContent' },
+                              { name: 'Flexible', value: 'flexible' },
+                              { name: 'Hidden', value: 'hidden' },
+                              { name: '1', value: '1' },
+                              { name: '2', value: '2' },
+                              { name: '3', value: '3' },
+                              { name: '4', value: '4' },
+                              { name: '5', value: '5' },
+                              { name: '6', value: '6' },
+                              { name: '7', value: '7' },
+                              { name: '8', value: '8' },
+                              { name: '9', value: '9' },
+                              { name: '10', value: '10' },
+                              { name: '11', value: '11' },
+                              { name: '12', value: '12' },
+                            ],
+                          },
+                        },
+                        {
+                          value: 'flexible',
+                          label: 'Column width (mobile)',
+                          key: 'columnWidthMobile',
+                          type: 'CUSTOM',
+                          configuration: {
+                            as: 'DROPDOWN',
+                            dataType: 'string',
+                            allowedInput: [
+                              { name: 'Fit content', value: 'fitContent' },
+                              { name: 'Flexible', value: 'flexible' },
+                              { name: 'Hidden', value: 'hidden' },
+                              { name: '1', value: '1' },
+                              { name: '2', value: '2' },
+                              { name: '3', value: '3' },
+                              { name: '4', value: '4' },
+                              { name: '5', value: '5' },
+                              { name: '6', value: '6' },
+                              { name: '7', value: '7' },
+                              { name: '8', value: '8' },
+                              { name: '9', value: '9' },
+                              { name: '10', value: '10' },
+                              { name: '11', value: '11' },
+                              { name: '12', value: '12' },
+                            ],
+                          },
+                        },
+                        {
+                          value: '',
+                          label: 'Height',
+                          key: 'columnHeight',
+                          type: 'TEXT',
+                          configuration: {
+                            as: 'UNIT',
+                          },
+                        },
+                        {
+                          value: 'White',
+                          label: 'Background color',
+                          key: 'backgroundColor',
+                          type: 'COLOR',
+                        },
+                        {
+                          type: 'CUSTOM',
+                          label: 'Horizontal Alignment',
+                          key: 'horizontalAlignment',
+                          value: 'inherit',
+                          configuration: {
+                            as: 'BUTTONGROUP',
+                            dataType: 'string',
+                            allowedInput: [
+                              { name: 'None', value: 'inherit' },
+                              { name: 'Left', value: 'flex-start' },
+                              { name: 'Center', value: 'center' },
+                              { name: 'Right', value: 'flex-end' },
+                            ],
+                          },
+                        },
+                        {
+                          type: 'CUSTOM',
+                          label: 'Vertical Alignment',
+                          key: 'verticalAlignment',
+                          value: 'inherit',
+                          configuration: {
+                            as: 'BUTTONGROUP',
+                            dataType: 'string',
+                            allowedInput: [
+                              { name: 'None', value: 'inherit' },
+                              { name: 'Top', value: 'flex-start' },
+                              { name: 'Center', value: 'center' },
+                              { name: 'Bottom', value: 'flex-end' },
+                            ],
+                          },
+                        },
+                        {
+                          value: ['S', 'S', 'S', 'S'],
+                          label: 'Outer space',
+                          key: 'outerSpacing',
+                          type: 'SIZES',
+                        },
+                        {
+                          value: ['M', 'M', 'M', 'M'],
+                          label: 'Inner space',
+                          key: 'innerSpacing',
+                          type: 'SIZES',
+                        },
+                        {
+                          value: false,
+                          label: 'Advanced settings',
+                          key: 'advancedSettings',
+                          type: 'TOGGLE',
+                        },
+                        {
+                          type: 'VARIABLE',
+                          label: 'Test attribute',
+                          key: 'dataComponentAttribute',
+                          value: ['Column'],
+                          configuration: {
+                            condition: {
+                              type: 'SHOW',
+                              option: 'advancedSettings',
+                              comparator: 'EQ',
+                              value: true,
+                            },
+                          },
+                        },
+                      ],
+                      descendants: [
+                        {
+                          name: 'DetailViewChild',
+                          options: [
+                            {
+                              value:
+                                property.kind === 'IMAGE'
+                                  ? ''
+                                  : enrichVarObj(property),
+                              label: 'Property',
+                              key: 'property',
+                              type: 'PROPERTY',
+                            },
+                            {
+                              type: 'VARIABLE',
+                              label: 'Label Text',
+                              key: 'labelText',
+                              value:
+                                property.kind === 'IMAGE'
+                                  ? [`${property.label}`]
+                                  : [],
+                            },
+                            {
+                              value: 'Body1',
+                              label: 'Type',
+                              key: 'type',
+                              type: 'FONT',
+                            },
+                            {
+                              type: 'CUSTOM',
+                              label: 'Text Alignment',
+                              key: 'textAlignment',
+                              value: 'left',
+                              configuration: {
+                                as: 'BUTTONGROUP',
+                                dataType: 'string',
+                                allowedInput: [
+                                  { name: 'Left', value: 'left' },
+                                  { name: 'Center', value: 'center' },
+                                  { name: 'Right', value: 'right' },
+                                ],
+                              },
+                            },
+                            {
+                              value: ['0rem', '0rem', '0rem', '0rem'],
+                              label: 'Outer space',
+                              key: 'outerSpacing',
+                              type: 'SIZES',
+                            },
+                            {
+                              value: false,
+                              label: 'Side by Side',
+                              key: 'sideBySide',
+                              type: 'TOGGLE',
+                            },
+                            {
+                              value: true,
+                              label: 'Styles',
+                              key: 'styles',
+                              type: 'TOGGLE',
+                            },
+                            {
+                              type: 'COLOR',
+                              label: 'Text color',
+                              key: 'textColor',
+                              value: 'Black',
+                              configuration: {
+                                condition: {
+                                  type: 'SHOW',
+                                  option: 'styles',
+                                  comparator: 'EQ',
+                                  value: true,
+                                },
+                              },
+                            },
+                            {
+                              type: 'CUSTOM',
+                              label: 'Font weight',
+                              key: 'fontWeight',
+                              value: '800',
+                              configuration: {
+                                as: 'DROPDOWN',
+                                dataType: 'string',
+                                allowedInput: [
+                                  { name: '100', value: '100' },
+                                  { name: '200', value: '200' },
+                                  { name: '300', value: '300' },
+                                  { name: '400', value: '400' },
+                                  { name: '500', value: '500' },
+                                  { name: '600', value: '600' },
+                                  { name: '700', value: '700' },
+                                  { name: '800', value: '800' },
+                                  { name: '900', value: '900' },
+                                ],
+                                condition: {
+                                  type: 'SHOW',
+                                  option: 'styles',
+                                  comparator: 'EQ',
+                                  value: true,
+                                },
+                              },
+                            },
+                            {
+                              value: false,
+                              label: 'Advanced settings',
+                              key: 'advancedSettings',
+                              type: 'TOGGLE',
+                            },
+                            {
+                              type: 'VARIABLE',
+                              label: 'Test attribute',
+                              key: 'dataComponentAttribute',
+                              value: ['DetailViewChild'],
+                              configuration: {
+                                condition: {
+                                  type: 'SHOW',
+                                  option: 'advancedSettings',
+                                  comparator: 'EQ',
+                                  value: true,
+                                },
+                              },
+                            },
+                          ],
+                          descendants:
+                            property.kind === 'IMAGE'
+                              ? [
+                                  {
+                                    name: 'Media',
+                                    options: [
+                                      {
+                                        label: 'Media type',
+                                        key: 'type',
+                                        value: 'img',
+                                        type: 'CUSTOM',
+                                        configuration: {
+                                          as: 'BUTTONGROUP',
+                                          dataType: 'string',
+                                          allowedInput: [
+                                            { name: 'Image', value: 'img' },
+                                            { name: 'Video', value: 'video' },
+                                            {
+                                              name: 'I-frame',
+                                              value: 'iframe',
+                                            },
+                                          ],
+                                        },
+                                      },
+                                      {
+                                        value: [property],
+                                        label: 'Source',
+                                        key: 'imageSource',
+                                        type: 'VARIABLE',
+                                        configuration: {
+                                          condition: {
+                                            type: 'SHOW',
+                                            option: 'type',
+                                            comparator: 'EQ',
+                                            value: 'img',
+                                          },
+                                        },
+                                      },
+                                      {
+                                        value: [],
+                                        label: 'Source',
+                                        key: 'videoSource',
+                                        type: 'VARIABLE',
+                                        configuration: {
+                                          condition: {
+                                            type: 'SHOW',
+                                            option: 'type',
+                                            comparator: 'EQ',
+                                            value: 'video',
+                                          },
+                                        },
+                                      },
+                                      {
+                                        value: [],
+                                        label: 'Source',
+                                        key: 'iframeSource',
+                                        type: 'VARIABLE',
+                                        configuration: {
+                                          condition: {
+                                            type: 'SHOW',
+                                            option: 'type',
+                                            comparator: 'EQ',
+                                            value: 'iframe',
+                                          },
+                                        },
+                                      },
+                                      {
+                                        type: 'CUSTOM',
+                                        label: 'Link to',
+                                        key: 'linkType',
+                                        value: 'internal',
+                                        configuration: {
+                                          as: 'BUTTONGROUP',
+                                          dataType: 'string',
+                                          allowedInput: [
+                                            {
+                                              name: 'Internal page',
+                                              value: 'internal',
+                                            },
+                                            {
+                                              name: 'External page',
+                                              value: 'external',
+                                            },
+                                          ],
+                                          condition: {
+                                            type: 'SHOW',
+                                            option: 'type',
+                                            comparator: 'EQ',
+                                            value: 'img',
+                                          },
+                                        },
+                                      },
+                                      {
+                                        value: '',
+                                        label: 'Page',
+                                        key: 'linkTo',
+                                        type: 'ENDPOINT',
+                                        configuration: {
+                                          condition: {
+                                            type: 'SHOW',
+                                            option: 'linkType',
+                                            comparator: 'EQ',
+                                            value: 'internal',
+                                          },
+                                        },
+                                      },
+                                      {
+                                        value: [''],
+                                        label: 'URL',
+                                        key: 'linkToExternal',
+                                        type: 'VARIABLE',
+                                        configuration: {
+                                          placeholder:
+                                            'Starts with https:// or http://',
+                                          condition: {
+                                            type: 'SHOW',
+                                            option: 'linkType',
+                                            comparator: 'EQ',
+                                            value: 'external',
+                                          },
+                                        },
+                                      },
+                                      {
+                                        value: [],
+                                        label: 'Image Alternative Text',
+                                        key: 'imgAlt',
+                                        type: 'VARIABLE',
+                                        configuration: {
+                                          condition: {
+                                            type: 'SHOW',
+                                            option: 'type',
+                                            comparator: 'EQ',
+                                            value: 'img',
+                                          },
+                                        },
+                                      },
+                                      {
+                                        value: [],
+                                        label: 'Title',
+                                        key: 'title',
+                                        type: 'VARIABLE',
+                                      },
+                                      {
+                                        type: 'SIZE',
+                                        label: 'Width',
+                                        key: 'width',
+                                        value: '100%',
+                                        configuration: {
+                                          as: 'UNIT',
+                                        },
+                                      },
+                                      {
+                                        type: 'SIZE',
+                                        label: 'Height',
+                                        key: 'height',
+                                        value: '',
+                                        configuration: {
+                                          as: 'UNIT',
+                                        },
+                                      },
+                                      {
+                                        value: ['0rem', '0rem', 'M', '0rem'],
+                                        label: 'Outer space',
+                                        key: 'outerSpacing',
+                                        type: 'SIZES',
+                                      },
+                                      {
+                                        value: false,
+                                        label: 'Advanced settings',
+                                        key: 'advancedSettings',
+                                        type: 'TOGGLE',
+                                      },
+                                      {
+                                        type: 'VARIABLE',
+                                        label: 'Test attribute',
+                                        key: 'dataComponentAttribute',
+                                        value: ['Media'],
+                                        configuration: {
+                                          condition: {
+                                            type: 'SHOW',
+                                            option: 'advancedSettings',
+                                            comparator: 'EQ',
+                                            value: true,
+                                          },
+                                        },
+                                      },
+                                    ],
+                                    descendants: [],
+                                  },
+                                ]
+                              : [],
+                        },
+                      ],
+                    },
+                  );
+                }
+              });
+            } else {
+              properties.forEach((property, index) => {
+                if (structure.descendants[0].type === 'COMPONENT')
+                  structure.descendants[0].descendants.push({
+                    name: 'Row',
+                    options: [
+                      {
+                        type: 'CUSTOM',
+                        label: 'Width',
+                        key: 'maxRowWidth',
+                        value: 'XL',
+                        configuration: {
+                          as: 'BUTTONGROUP',
+                          dataType: 'string',
+                          allowedInput: [
+                            { name: 'S', value: 'S' },
+                            { name: 'M', value: 'M' },
+                            { name: 'L', value: 'L' },
+                            { name: 'XL', value: 'XL' },
+                            { name: 'Full', value: 'Full' },
+                          ],
+                        },
+                      },
+                      {
+                        value: '',
+                        label: 'Height',
+                        key: 'rowHeight',
+                        type: 'TEXT',
+                        configuration: {
+                          as: 'UNIT',
+                        },
+                      },
+                      {
+                        value: 'Accent1',
+                        label: 'Background color',
+                        key: 'backgroundColor',
+                        type: 'COLOR',
+                      },
+                      {
+                        value: [
+                          '0rem',
+                          '0rem',
+                          `${properties.length - 1 === index ? '0rem' : 'M'}`,
+                          '0rem',
+                        ],
+                        label: 'Outer space',
+                        key: 'outerSpacing',
+                        type: 'SIZES',
+                      },
+                      {
+                        value: false,
+                        label: 'Advanced settings',
+                        key: 'advancedSettings',
+                        type: 'TOGGLE',
+                      },
+                      {
+                        type: 'VARIABLE',
+                        label: 'Test attribute',
+                        key: 'dataComponentAttribute',
+                        value: ['Row'],
+                        configuration: {
+                          condition: {
+                            type: 'SHOW',
+                            option: 'advancedSettings',
+                            comparator: 'EQ',
+                            value: true,
+                          },
+                        },
+                      },
+                    ],
+                    descendants: [
+                      {
+                        name: 'Column',
+                        options: [
+                          {
+                            label: 'Toggle visibility',
+                            key: 'visible',
+                            value: true,
+                            type: 'TOGGLE',
+                            configuration: {
+                              as: 'VISIBILITY',
+                            },
+                          },
+                          {
+                            value: '3',
+                            label: 'Column width',
+                            key: 'columnWidth',
+                            type: 'CUSTOM',
+                            configuration: {
+                              as: 'DROPDOWN',
+                              dataType: 'string',
+                              allowedInput: [
+                                { name: 'Fit content', value: 'fitContent' },
+                                { name: 'Flexible', value: 'flexible' },
+                                { name: 'Hidden', value: 'hidden' },
+                                { name: '1', value: '1' },
+                                { name: '2', value: '2' },
+                                { name: '3', value: '3' },
+                                { name: '4', value: '4' },
+                                { name: '5', value: '5' },
+                                { name: '6', value: '6' },
+                                { name: '7', value: '7' },
+                                { name: '8', value: '8' },
+                                { name: '9', value: '9' },
+                                { name: '10', value: '10' },
+                                { name: '11', value: '11' },
+                                { name: '12', value: '12' },
+                              ],
+                            },
+                          },
+                          {
+                            value: '3',
+                            label: 'Column width (tablet landscape)',
+                            key: 'columnWidthTabletLandscape',
+                            type: 'CUSTOM',
+                            configuration: {
+                              as: 'DROPDOWN',
+                              dataType: 'string',
+                              allowedInput: [
+                                { name: 'Fit content', value: 'fitContent' },
+                                { name: 'Flexible', value: 'flexible' },
+                                { name: 'Hidden', value: 'hidden' },
+                                { name: '1', value: '1' },
+                                { name: '2', value: '2' },
+                                { name: '3', value: '3' },
+                                { name: '4', value: '4' },
+                                { name: '5', value: '5' },
+                                { name: '6', value: '6' },
+                                { name: '7', value: '7' },
+                                { name: '8', value: '8' },
+                                { name: '9', value: '9' },
+                                { name: '10', value: '10' },
+                                { name: '11', value: '11' },
+                                { name: '12', value: '12' },
+                              ],
+                            },
+                          },
+                          {
+                            value: 'flexible',
+                            label: 'Column width (tablet portrait)',
+                            key: 'columnWidthTabletPortrait',
+                            type: 'CUSTOM',
+                            configuration: {
+                              as: 'DROPDOWN',
+                              dataType: 'string',
+                              allowedInput: [
+                                { name: 'Fit content', value: 'fitContent' },
+                                { name: 'Flexible', value: 'flexible' },
+                                { name: 'Hidden', value: 'hidden' },
+                                { name: '1', value: '1' },
+                                { name: '2', value: '2' },
+                                { name: '3', value: '3' },
+                                { name: '4', value: '4' },
+                                { name: '5', value: '5' },
+                                { name: '6', value: '6' },
+                                { name: '7', value: '7' },
+                                { name: '8', value: '8' },
+                                { name: '9', value: '9' },
+                                { name: '10', value: '10' },
+                                { name: '11', value: '11' },
+                                { name: '12', value: '12' },
+                              ],
+                            },
+                          },
+                          {
+                            value: 'flexible',
+                            label: 'Column width (mobile)',
+                            key: 'columnWidthMobile',
+                            type: 'CUSTOM',
+                            configuration: {
+                              as: 'DROPDOWN',
+                              dataType: 'string',
+                              allowedInput: [
+                                { name: 'Fit content', value: 'fitContent' },
+                                { name: 'Flexible', value: 'flexible' },
+                                { name: 'Hidden', value: 'hidden' },
+                                { name: '1', value: '1' },
+                                { name: '2', value: '2' },
+                                { name: '3', value: '3' },
+                                { name: '4', value: '4' },
+                                { name: '5', value: '5' },
+                                { name: '6', value: '6' },
+                                { name: '7', value: '7' },
+                                { name: '8', value: '8' },
+                                { name: '9', value: '9' },
+                                { name: '10', value: '10' },
+                                { name: '11', value: '11' },
+                                { name: '12', value: '12' },
+                              ],
+                            },
+                          },
+                          {
+                            value: '',
+                            label: 'Height',
+                            key: 'columnHeight',
+                            type: 'TEXT',
+                            configuration: {
+                              as: 'UNIT',
+                            },
+                          },
+                          {
+                            value: 'White',
+                            label: 'Background color',
+                            key: 'backgroundColor',
+                            type: 'COLOR',
+                          },
+                          {
+                            type: 'CUSTOM',
+                            label: 'Horizontal Alignment',
+                            key: 'horizontalAlignment',
+                            value: 'flex-end',
+                            configuration: {
+                              as: 'BUTTONGROUP',
+                              dataType: 'string',
+                              allowedInput: [
+                                { name: 'None', value: 'inherit' },
+                                { name: 'Left', value: 'flex-start' },
+                                { name: 'Center', value: 'center' },
+                                { name: 'Right', value: 'flex-end' },
+                              ],
+                            },
+                          },
+                          {
+                            type: 'CUSTOM',
+                            label: 'Vertical Alignment',
+                            key: 'verticalAlignment',
+                            value: 'inherit',
+                            configuration: {
+                              as: 'BUTTONGROUP',
+                              dataType: 'string',
+                              allowedInput: [
+                                { name: 'None', value: 'inherit' },
+                                { name: 'Top', value: 'flex-start' },
+                                { name: 'Center', value: 'center' },
+                                { name: 'Bottom', value: 'flex-end' },
+                              ],
+                            },
+                          },
+                          {
+                            value: ['0rem', '0rem', '0rem', '0rem'],
+                            label: 'Outer space',
+                            key: 'outerSpacing',
+                            type: 'SIZES',
+                          },
+                          {
+                            value: ['0rem', '0rem', '0rem', '0rem'],
+                            label: 'Inner space',
+                            key: 'innerSpacing',
+                            type: 'SIZES',
+                          },
+                          {
+                            value: false,
+                            label: 'Advanced settings',
+                            key: 'advancedSettings',
+                            type: 'TOGGLE',
+                          },
+                          {
+                            type: 'VARIABLE',
+                            label: 'Test attribute',
+                            key: 'dataComponentAttribute',
+                            value: ['Column'],
+                            configuration: {
+                              condition: {
+                                type: 'SHOW',
+                                option: 'advancedSettings',
+                                comparator: 'EQ',
+                                value: true,
+                              },
+                            },
+                          },
+                        ],
+                        descendants: [
+                          {
+                            name: 'Text',
+                            options: [
+                              {
+                                type: 'VARIABLE',
+                                label: 'Content',
+                                key: 'content',
+                                value: [`${property.label}: `],
+                                configuration: {
+                                  as: 'MULTILINE',
+                                },
+                              },
+                              {
+                                type: 'TOGGLE',
+                                label: 'Display Rich Text',
+                                key: 'useInnerHtml',
+                                value: false,
+                              },
+                              {
+                                value: 'Body1',
+                                label: 'Type',
+                                key: 'type',
+                                type: 'FONT',
+                              },
+                              {
+                                type: 'CUSTOM',
+                                label: 'Text Alignment',
+                                key: 'textAlignment',
+                                value: 'left',
+                                configuration: {
+                                  as: 'BUTTONGROUP',
+                                  dataType: 'string',
+                                  allowedInput: [
+                                    { name: 'Left', value: 'left' },
+                                    { name: 'Center', value: 'center' },
+                                    { name: 'Right', value: 'right' },
+                                  ],
+                                },
+                              },
+                              {
+                                value: ['0rem', 'S', '0rem', '0rem'],
+                                label: 'Outer space',
+                                key: 'outerSpacing',
+                                type: 'SIZES',
+                              },
+                              {
+                                type: 'CUSTOM',
+                                label: 'Link to',
+                                key: 'linkType',
+                                value: 'internal',
+                                configuration: {
+                                  as: 'BUTTONGROUP',
+                                  dataType: 'string',
+                                  allowedInput: [
+                                    {
+                                      name: 'Internal page',
+                                      value: 'internal',
+                                    },
+                                    {
+                                      name: 'External page',
+                                      value: 'external',
+                                    },
+                                  ],
+                                },
+                              },
+                              {
+                                value: '',
+                                label: 'Page',
+                                key: 'linkTo',
+                                type: 'ENDPOINT',
+                                configuration: {
+                                  condition: {
+                                    type: 'SHOW',
+                                    option: 'linkType',
+                                    comparator: 'EQ',
+                                    value: 'internal',
+                                  },
+                                },
+                              },
+                              {
+                                value: [''],
+                                label: 'URL',
+                                key: 'linkToExternal',
+                                type: 'VARIABLE',
+                                configuration: {
+                                  placeholder:
+                                    'Starts with https:// or http://',
+                                  condition: {
+                                    type: 'SHOW',
+                                    option: 'linkType',
+                                    comparator: 'EQ',
+                                    value: 'external',
+                                  },
+                                },
+                              },
+                              {
+                                value: false,
+                                label: 'Styles',
+                                key: 'styles',
+                                type: 'TOGGLE',
+                              },
+                              {
+                                type: 'COLOR',
+                                label: 'Text color',
+                                key: 'textColor',
+                                value: 'Black',
+                                configuration: {
+                                  condition: {
+                                    type: 'SHOW',
+                                    option: 'styles',
+                                    comparator: 'EQ',
+                                    value: true,
+                                  },
+                                },
+                              },
+                              {
+                                type: 'CUSTOM',
+                                label: 'Font weight',
+                                key: 'fontWeight',
+                                value: '400',
+                                configuration: {
+                                  as: 'DROPDOWN',
+                                  dataType: 'string',
+                                  allowedInput: [
+                                    { name: '100', value: '100' },
+                                    { name: '200', value: '200' },
+                                    { name: '300', value: '300' },
+                                    { name: '400', value: '400' },
+                                    { name: '500', value: '500' },
+                                    { name: '600', value: '600' },
+                                    { name: '700', value: '700' },
+                                    { name: '800', value: '800' },
+                                    { name: '900', value: '900' },
+                                  ],
+                                  condition: {
+                                    type: 'SHOW',
+                                    option: 'styles',
+                                    comparator: 'EQ',
+                                    value: true,
+                                  },
+                                },
+                              },
+                              {
+                                value: false,
+                                label: 'Advanced settings',
+                                key: 'advancedSettings',
+                                type: 'TOGGLE',
+                              },
+                              {
+                                type: 'VARIABLE',
+                                label: 'Test attribute',
+                                key: 'dataComponentAttribute',
+                                value: ['Text'],
+                                configuration: {
+                                  condition: {
+                                    type: 'SHOW',
+                                    option: 'advancedSettings',
+                                    comparator: 'EQ',
+                                    value: true,
+                                  },
+                                },
+                              },
+                            ],
+                            descendants: [],
+                          },
+                        ],
+                      },
+                      {
+                        name: 'Column',
+                        options: [
+                          {
+                            label: 'Toggle visibility',
+                            key: 'visible',
+                            value: true,
+                            type: 'TOGGLE',
+                            configuration: {
+                              as: 'VISIBILITY',
+                            },
+                          },
+                          {
+                            value: '9',
+                            label: 'Column width',
+                            key: 'columnWidth',
+                            type: 'CUSTOM',
+                            configuration: {
+                              as: 'DROPDOWN',
+                              dataType: 'string',
+                              allowedInput: [
+                                { name: 'Fit content', value: 'fitContent' },
+                                { name: 'Flexible', value: 'flexible' },
+                                { name: 'Hidden', value: 'hidden' },
+                                { name: '1', value: '1' },
+                                { name: '2', value: '2' },
+                                { name: '3', value: '3' },
+                                { name: '4', value: '4' },
+                                { name: '5', value: '5' },
+                                { name: '6', value: '6' },
+                                { name: '7', value: '7' },
+                                { name: '8', value: '8' },
+                                { name: '9', value: '9' },
+                                { name: '10', value: '10' },
+                                { name: '11', value: '11' },
+                                { name: '12', value: '12' },
+                              ],
+                            },
+                          },
+                          {
+                            value: '9',
+                            label: 'Column width (tablet landscape)',
+                            key: 'columnWidthTabletLandscape',
+                            type: 'CUSTOM',
+                            configuration: {
+                              as: 'DROPDOWN',
+                              dataType: 'string',
+                              allowedInput: [
+                                { name: 'Fit content', value: 'fitContent' },
+                                { name: 'Flexible', value: 'flexible' },
+                                { name: 'Hidden', value: 'hidden' },
+                                { name: '1', value: '1' },
+                                { name: '2', value: '2' },
+                                { name: '3', value: '3' },
+                                { name: '4', value: '4' },
+                                { name: '5', value: '5' },
+                                { name: '6', value: '6' },
+                                { name: '7', value: '7' },
+                                { name: '8', value: '8' },
+                                { name: '9', value: '9' },
+                                { name: '10', value: '10' },
+                                { name: '11', value: '11' },
+                                { name: '12', value: '12' },
+                              ],
+                            },
+                          },
+                          {
+                            value: 'flexible',
+                            label: 'Column width (tablet portrait)',
+                            key: 'columnWidthTabletPortrait',
+                            type: 'CUSTOM',
+                            configuration: {
+                              as: 'DROPDOWN',
+                              dataType: 'string',
+                              allowedInput: [
+                                { name: 'Fit content', value: 'fitContent' },
+                                { name: 'Flexible', value: 'flexible' },
+                                { name: 'Hidden', value: 'hidden' },
+                                { name: '1', value: '1' },
+                                { name: '2', value: '2' },
+                                { name: '3', value: '3' },
+                                { name: '4', value: '4' },
+                                { name: '5', value: '5' },
+                                { name: '6', value: '6' },
+                                { name: '7', value: '7' },
+                                { name: '8', value: '8' },
+                                { name: '9', value: '9' },
+                                { name: '10', value: '10' },
+                                { name: '11', value: '11' },
+                                { name: '12', value: '12' },
+                              ],
+                            },
+                          },
+                          {
+                            value: 'flexible',
+                            label: 'Column width (mobile)',
+                            key: 'columnWidthMobile',
+                            type: 'CUSTOM',
+                            configuration: {
+                              as: 'DROPDOWN',
+                              dataType: 'string',
+                              allowedInput: [
+                                { name: 'Fit content', value: 'fitContent' },
+                                { name: 'Flexible', value: 'flexible' },
+                                { name: 'Hidden', value: 'hidden' },
+                                { name: '1', value: '1' },
+                                { name: '2', value: '2' },
+                                { name: '3', value: '3' },
+                                { name: '4', value: '4' },
+                                { name: '5', value: '5' },
+                                { name: '6', value: '6' },
+                                { name: '7', value: '7' },
+                                { name: '8', value: '8' },
+                                { name: '9', value: '9' },
+                                { name: '10', value: '10' },
+                                { name: '11', value: '11' },
+                                { name: '12', value: '12' },
+                              ],
+                            },
+                          },
+                          {
+                            value: '',
+                            label: 'Height',
+                            key: 'columnHeight',
+                            type: 'TEXT',
+                            configuration: {
+                              as: 'UNIT',
+                            },
+                          },
+                          {
+                            value: 'White',
+                            label: 'Background color',
+                            key: 'backgroundColor',
+                            type: 'COLOR',
+                          },
+                          {
+                            type: 'CUSTOM',
+                            label: 'Horizontal Alignment',
+                            key: 'horizontalAlignment',
+                            value: 'inherit',
+                            configuration: {
+                              as: 'BUTTONGROUP',
+                              dataType: 'string',
+                              allowedInput: [
+                                { name: 'None', value: 'inherit' },
+                                { name: 'Left', value: 'flex-start' },
+                                { name: 'Center', value: 'center' },
+                                { name: 'Right', value: 'flex-end' },
+                              ],
+                            },
+                          },
+                          {
+                            type: 'CUSTOM',
+                            label: 'Vertical Alignment',
+                            key: 'verticalAlignment',
+                            value: 'inherit',
+                            configuration: {
+                              as: 'BUTTONGROUP',
+                              dataType: 'string',
+                              allowedInput: [
+                                { name: 'None', value: 'inherit' },
+                                { name: 'Top', value: 'flex-start' },
+                                { name: 'Center', value: 'center' },
+                                { name: 'Bottom', value: 'flex-end' },
+                              ],
+                            },
+                          },
+                          {
+                            value: ['0rem', '0rem', '0rem', '0rem'],
+                            label: 'Outer space',
+                            key: 'outerSpacing',
+                            type: 'SIZES',
+                          },
+                          {
+                            value: ['0rem', '0rem', '0rem', '0rem'],
+                            label: 'Inner space',
+                            key: 'innerSpacing',
+                            type: 'SIZES',
+                          },
+                          {
+                            value: false,
+                            label: 'Advanced settings',
+                            key: 'advancedSettings',
+                            type: 'TOGGLE',
+                          },
+                          {
+                            type: 'VARIABLE',
+                            label: 'Test attribute',
+                            key: 'dataComponentAttribute',
+                            value: ['Column'],
+                            configuration: {
+                              condition: {
+                                type: 'SHOW',
+                                option: 'advancedSettings',
+                                comparator: 'EQ',
+                                value: true,
+                              },
+                            },
+                          },
+                        ],
+                        descendants: [
+                          {
+                            name: 'Text',
+                            options: [
+                              {
+                                type: 'VARIABLE',
+                                label: 'Content',
+                                key: 'content',
+                                value: [enrichVarObj(property)],
+                                configuration: {
+                                  as: 'MULTILINE',
+                                },
+                              },
+                              {
+                                type: 'TOGGLE',
+                                label: 'Display Rich Text',
+                                key: 'useInnerHtml',
+                                value: false,
+                              },
+                              {
+                                value: 'Body1',
+                                label: 'Type',
+                                key: 'type',
+                                type: 'FONT',
+                              },
+                              {
+                                type: 'CUSTOM',
+                                label: 'Text Alignment',
+                                key: 'textAlignment',
+                                value: 'left',
+                                configuration: {
+                                  as: 'BUTTONGROUP',
+                                  dataType: 'string',
+                                  allowedInput: [
+                                    { name: 'Left', value: 'left' },
+                                    { name: 'Center', value: 'center' },
+                                    { name: 'Right', value: 'right' },
+                                  ],
+                                },
+                              },
+                              {
+                                value: ['0rem', '0rem', '0rem', '0rem'],
+                                label: 'Outer space',
+                                key: 'outerSpacing',
+                                type: 'SIZES',
+                              },
+                              {
+                                type: 'CUSTOM',
+                                label: 'Link to',
+                                key: 'linkType',
+                                value: 'internal',
+                                configuration: {
+                                  as: 'BUTTONGROUP',
+                                  dataType: 'string',
+                                  allowedInput: [
+                                    {
+                                      name: 'Internal page',
+                                      value: 'internal',
+                                    },
+                                    {
+                                      name: 'External page',
+                                      value: 'external',
+                                    },
+                                  ],
+                                },
+                              },
+                              {
+                                value: '',
+                                label: 'Page',
+                                key: 'linkTo',
+                                type: 'ENDPOINT',
+                                configuration: {
+                                  condition: {
+                                    type: 'SHOW',
+                                    option: 'linkType',
+                                    comparator: 'EQ',
+                                    value: 'internal',
+                                  },
+                                },
+                              },
+                              {
+                                value: [''],
+                                label: 'URL',
+                                key: 'linkToExternal',
+                                type: 'VARIABLE',
+                                configuration: {
+                                  placeholder:
+                                    'Starts with https:// or http://',
+                                  condition: {
+                                    type: 'SHOW',
+                                    option: 'linkType',
+                                    comparator: 'EQ',
+                                    value: 'external',
+                                  },
+                                },
+                              },
+                              {
+                                value: false,
+                                label: 'Styles',
+                                key: 'styles',
+                                type: 'TOGGLE',
+                              },
+                              {
+                                type: 'COLOR',
+                                label: 'Text color',
+                                key: 'textColor',
+                                value: 'Black',
+                                configuration: {
+                                  condition: {
+                                    type: 'SHOW',
+                                    option: 'styles',
+                                    comparator: 'EQ',
+                                    value: true,
+                                  },
+                                },
+                              },
+                              {
+                                type: 'CUSTOM',
+                                label: 'Font weight',
+                                key: 'fontWeight',
+                                value: '400',
+                                configuration: {
+                                  as: 'DROPDOWN',
+                                  dataType: 'string',
+                                  allowedInput: [
+                                    { name: '100', value: '100' },
+                                    { name: '200', value: '200' },
+                                    { name: '300', value: '300' },
+                                    { name: '400', value: '400' },
+                                    { name: '500', value: '500' },
+                                    { name: '600', value: '600' },
+                                    { name: '700', value: '700' },
+                                    { name: '800', value: '800' },
+                                    { name: '900', value: '900' },
+                                  ],
+                                  condition: {
+                                    type: 'SHOW',
+                                    option: 'styles',
+                                    comparator: 'EQ',
+                                    value: true,
+                                  },
+                                },
+                              },
+                              {
+                                value: false,
+                                label: 'Advanced settings',
+                                key: 'advancedSettings',
+                                type: 'TOGGLE',
+                              },
+                              {
+                                type: 'VARIABLE',
+                                label: 'Test attribute',
+                                key: 'dataComponentAttribute',
+                                value: ['Text'],
+                                configuration: {
+                                  condition: {
+                                    type: 'SHOW',
+                                    option: 'advancedSettings',
+                                    comparator: 'EQ',
+                                    value: true,
+                                  },
+                                },
+                              },
+                            ],
+                            descendants: [],
+                          },
+                        ],
+                      },
+                    ],
+                  });
+              });
             }
           }
+          save(newPrefab);
         }}
         onClose={close}
       />
@@ -190,7 +1512,7 @@ const attrs = {
   variables: [],
 };
 
-export default makePrefab('Detail view (TS)', attrs, beforeCreate, [
+export default makePrefab('Detail view', attrs, beforeCreate, [
   DataContainer({}, [
     Box(
       {
@@ -218,29 +1540,7 @@ export default makePrefab('Detail view (TS)', attrs, beforeCreate, [
               }),
             },
           },
-          [
-            Column(
-              {
-                options: {
-                  ...columnOptions,
-                  outerSpacing: sizes('Outer space', {
-                    value: ['S', 'S', 'S', 'S'],
-                  }),
-                  backgroundColor: color('Background color', {
-                    value: ThemeColor.WHITE,
-                  }),
-                },
-              },
-              [
-                DetailViewChild(
-                  {
-                    ref: { id: '#detailViewChild' },
-                  },
-                  [],
-                ),
-              ],
-            ),
-          ],
+          [],
         ),
       ],
     ),
