+++ src/prefabs/detailViewChild.tsx
@@ -1,4 +1,6 @@
-import { Icon, prefab } from '@betty-blocks/component-sdk';
+import { color, Icon, prefab, ThemeColor } from '@betty-blocks/component-sdk';
+import { Column } from './structures/Column';
+import { options as columnOptions } from './structures/Column/options';
 import { DetailViewChild } from './structures/DetailViewChild';
 
 const attrs = {
@@ -7,6 +9,16 @@ const attrs = {
   category: 'DATA',
 };
 
-export default prefab('Detail view child (TS)', attrs, undefined, [
-  DetailViewChild({}, []),
+export default prefab('Detail view child', attrs, undefined, [
+  Column(
+    {
+      options: {
+        ...columnOptions,
+        backgroundColor: color('Background color', {
+          value: ThemeColor.WHITE,
+        }),
+      },
+    },
+    [DetailViewChild({}, [])],
+  ),
 ]);
