+++ src/prefabs/structures/Column/index.ts
@@ -1,21 +1,26 @@
 import {
   component,
   PrefabComponentOption,
-  PrefabComponent,
+  PrefabReference,
+  PrefabComponentStyle,
 } from '@betty-blocks/component-sdk';
 import { options as defaults } from './options';
 
 type OptionProducer = (key: string) => PrefabComponentOption;
 
 export interface Configuration {
-  label?: string;
   options?: Record<string, OptionProducer>;
+  style?: PrefabComponentStyle;
+  ref?: { id: string };
 }
 
 export const Column = (
   config: Configuration,
-  descendants: PrefabComponent[] = [],
-): PrefabComponent => {
+  descendants: PrefabReference[] = [],
+) => {
   const options = { ...(config.options || defaults) };
-  return component('Column', { options }, descendants);
+  const style = { ...config.style };
+  const ref = config.ref ? { ...config.ref } : undefined;
+
+  return component('Column', { options, style, ref }, descendants);
 };
