+++ src/prefabs/structures/Media/index.ts
@@ -4,11 +4,13 @@ import {
   OptionProducer,
   PrefabComponentStyle,
 } from '@betty-blocks/component-sdk';
+
 import { options as defaults } from './options';
 
 export interface Configuration {
   options?: Record<string, OptionProducer>;
   style?: PrefabComponentStyle;
+  ref?: { id: string };
 }
 
 export const Media = (
@@ -17,6 +19,7 @@ export const Media = (
 ) => {
   const options = { ...(config.options || defaults) };
   const style = { ...config.style };
+  const ref = config.ref ? { ...config.ref } : undefined;
 
-  return component('Media', { options, style }, descendants);
+  return component('Media', { options, style, ref }, descendants);
 };
